yolov10_requirements.txt
aiofiles==23.2.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cmake==3.30.2
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
dlib @ file:///home/conda/feedstock_root/build_artifacts/dlib-split_1725133999133/work
exceptiongroup==1.2.2
fastapi==0.112.2
ffmpy==0.4.0
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
GPUtil==1.4.0
gradio==4.31.5
gradio_client==0.16.4
graphsurgeon @ file:///home/ymm/Softwares/TensorRT-8.6.1.6/graphsurgeon/graphsurgeon-0.4.6-py2.py3-none-any.whl#sha256=0fbadaefbbe6e9920b9f814ae961c4a279be602812edf3ed7fb9cc6f8f4809fe
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.2
humanfriendly==10.0
idna==3.8
importlib_resources==6.4.4
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lit==18.1.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.5.5
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.4.2.65
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.4.99
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.4.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.4.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.0.44
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.5.119
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.6.0.99
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.3.0.142
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.4.99
onnx==1.14.0
onnxruntime==1.15.1
onnxruntime-gpu==1.18.0
onnxslim==0.1.31
opencv-python==4.8.0.76
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.27.3
psutil==5.9.8
py-cpuinfo==9.0.0
pycocotools==2.0.7
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
ruff==0.6.2
safetensors==0.4.3
scipy==1.13.0
seaborn==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
shiboken6==6.7.2
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
sympy==1.13.2
tensorrt @ file:///home/ymm/Softwares/TensorRT-8.6.1.6/python/tensorrt-8.6.1-cp310-none-linux_x86_64.whl#sha256=2684b4772cb16088184266728a0668f5dac14e66f088c4ccff2096ccb222d74c
thop==0.1.1.post2209072238
tomlkit==0.12.0
torch==2.0.1
torchaudio==2.4.1+cu124
torchvision==0.15.2
tqdm==4.66.5
triton==2.0.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
uff @ file:///home/ymm/Softwares/TensorRT-8.6.1.6/uff/uff-0.6.9-py2.py3-none-any.whl#sha256=618a3f812d491f0d3c4f2e38b99e03217ca37b206db14cee079f2bf681eb4fe3
ultralytics==8.1.34
ultralytics-thop==2.0.8
urllib3==2.2.2
uvicorn==0.30.6
websockets==11.0.3
